+++ generator/page_generators/devlog/new_systems_requires_major_changes.js
@@ -35,7 +35,7 @@ page.paragraph(`The plan was, during semantic analysis, to make a first pass and
 
 page.paragraph("I had met all the requirements for OID that I had laid out for myself, so I went to implement it. However, when I began considering the exact implementation details, I realized that the plan could cause lots of waiting during multithreading and tank performance. To explain how, consider the following code:");
 
-page.code_block(Language.Panther, `def some_lib = @import("some_lib.pthr");
+page.code_block(Language.PANTHER, `def some_lib = @import("some_lib.pthr");
 
 func bar = () #pub -> Foo {
 	// do something...