+++ generator/page_generators/documentation/panther/literals.js
@@ -64,7 +64,7 @@ page.bullets([
 	`${html.inline_code("0x")} followed by 1 or more hexadecimal digits (${nums[0]}, ${nums[1]}, ${nums[2]}, ${nums[3]}, ${nums[4]}, ${nums[5]}, ${nums[6]}, ${nums[7]}, ${nums[8]}, ${nums[9]}, ${nums['a']}, ${nums['A']}, ${nums['b']}, ${nums['B']}, ${nums['c']}, ${nums['C']}, ${nums['d']}, ${nums['D']}, ${nums['e']}, ${nums['E']}, ${nums['f']}, ${nums['F']})`,
 ]);
 page.text(`An integer literal may optionally have a suffix of scientific notation which is: an ${html.inline_code("E")} or an ${html.inline_code("e")}, followed by an optional ${html.inline_code("+")} or ${html.inline_code("-")} (omitting is equivalent to ${html.inline_code("+")}), followed by 1 or more decimal digits (${nums[0]}, ${nums[1]}, ${nums[2]}, ${nums[3]}, ${nums[4]}, ${nums[5]}, ${nums[6]}, ${nums[7]}, ${nums[8]}, ${nums[9]}).`);
-page.text(`Integer literals may optionally be broken up by underscores ( ${html.inline_code('_')} ) anywhere except for the beginning. This means that ${html.inline_code(syntax_highlighting.panther("12_345_"))} is equivalent to ${html.inline_code(syntax_highlighting.panther("12345"))}.`);
+page.text(`Integer literals may optionally be broken up by underscores ( ${html.inline_code('_')} ) anywhere except for the beginning. This means that ${page.inline_code_block(Language.PANTHER, "12_345_")} is equivalent to ${page.inline_code_block(Language.PANTHER, "12345")}.`);
 
 
 page.h2Searchable("Float Literals", "float-literals");
@@ -73,20 +73,20 @@ page.bullets([
 	`1 or more decimal digits (${nums[0]}, ${nums[1]}, ${nums[2]}, ${nums[3]}, ${nums[4]}, ${nums[5]}, ${nums[6]}, ${nums[7]}, ${nums[8]}, ${nums[9]}), followed by  followed a period (${html.inline_code('.')}), followed by 1 or more decimal digits (${nums[0]}, ${nums[1]}, ${nums[2]}, ${nums[3]}, ${nums[4]}, ${nums[5]}, ${nums[6]}, ${nums[7]}, ${nums[8]}, ${nums[9]}). This may optionally be have suffix of scientific notation which is: an ${html.inline_code("E")} or an ${html.inline_code("e")}, followed by an optional ${html.inline_code("+")} or ${html.inline_code("-")} (omitting is equivalent to ${html.inline_code("+")}), followed by 1 or more decimal digits (${nums[0]}, ${nums[1]}, ${nums[2]}, ${nums[3]}, ${nums[4]}, ${nums[5]}, ${nums[6]}, ${nums[7]}, ${nums[8]}, ${nums[9]}).`,
 	`${html.inline_code("0x")}, followed by 1 or more hexadecimal digits (${nums[0]}, ${nums[1]}, ${nums[2]}, ${nums[3]}, ${nums[4]}, ${nums[5]}, ${nums[6]}, ${nums[7]}, ${nums[8]}, ${nums[9]}, ${nums['a']}, ${nums['A']}, ${nums['b']}, ${nums['B']}, ${nums['c']}, ${nums['C']}, ${nums['d']}, ${nums['D']}, ${nums['e']}, ${nums['E']}, ${nums['f']}, ${nums['F']}), followed by a period (${html.inline_code('.')}), followed by 1 or more hexadecimal digits (${nums[0]}, ${nums[1]}, ${nums[2]}, ${nums[3]}, ${nums[4]}, ${nums[5]}, ${nums[6]}, ${nums[7]}, ${nums[8]}, ${nums[9]}, ${nums['a']}, ${nums['A']}, ${nums['b']}, ${nums['B']}, ${nums['c']}, ${nums['C']}, ${nums['d']}, ${nums['D']}, ${nums['e']}, ${nums['E']}, ${nums['f']}, ${nums['F']})`,
 ]);
-page.text(`Float literals may optionally be broken up by underscores ( ${html.inline_code('_')} ) anywhere except for the beginning. This means that ${html.inline_code(syntax_highlighting.panther("1_234.5"))} is equivalent to ${html.inline_code(syntax_highlighting.panther("1234.5"))}.`);
+page.text(`Float literals may optionally be broken up by underscores ( ${html.inline_code('_')} ) anywhere except for the beginning. This means that ${page.inline_code_block(Language.PANTHER, "1_234.5")} is equivalent to ${page.inline_code_block(Language.PANTHER, "1234.5")}.`);
 
 
 
 page.h2Searchable("Boolean Literals", "boolean-literals");
-page.text(`Boolean literals are expressions represent a single boolean value. They are type ${html.inline_code(syntax_highlighting.panther("Bool"))} and are represented by the following:`);
+page.text(`Boolean literals are expressions represent a single boolean value. They are type ${page.inline_code_block(Language.PANTHER, "Bool")} and are represented by the following:`);
 page.bullets([
-	html.inline_code(syntax_highlighting.panther("true")),
-	html.inline_code(syntax_highlighting.panther("false")),
+	page.inline_code_block(Language.PANTHER, "true"),
+	page.inline_code_block(Language.PANTHER, "false"),
 ]);
 
 
 page.h2Searchable("Character Literals", "character-literals");
-page.text(`Character literals are expressions that represent a single character and are represented as ASCII. They are type ${html.inline_code(syntax_highlighting.panther("Char"))}. They are represented by two single-quote characters (${html.inline_code('\'')}) containing one the following:`);
+page.text(`Character literals are expressions that represent a single character and are represented as ASCII. They are type ${page.inline_code_block(Language.PANTHER, "Char")}. They are represented by two single-quote characters (${html.inline_code('\'')}) containing one the following:`);
 page.bullets([
 	`Any basic ASCII printable character excluding ${html.inline_code('\'')}, ${html.inline_code("\\")}, new-line, or carriage return`,
 	`A ${terms.get("character escape code")}`
@@ -94,19 +94,19 @@ page.bullets([
 
 
 page.h2Searchable("String Literals", "string-literals");
-page.text(`String literals are expressions that represent a string of text. They are type ${html.inline_code(syntax_highlighting.panther("[Char:N;'\\0']"))} where ${html.inline_code("N")} is the number of characters in the string literal. They are represented by two double-quotes characters (${html.inline_code('"')}) containing 0 or more of the following:`);
+page.text(`String literals are expressions that represent a string of text. They are type ${page.inline_code_block(Language.PANTHER, "[Char:N;'\\0']")} where ${html.inline_code("N")} is the number of characters in the string literal. They are represented by two double-quotes characters (${html.inline_code('"')}) containing 0 or more of the following:`);
 page.bullets([
 	`Any basic ASCII printable character excluding ${html.inline_code('"')}, and ${html.inline_code("\\")}`,
 	`A ${terms.get("character escape code")}`
 ]);
 
 page.h3("Notes");
-page.text(`As suggested by the type, string literals automatically have a null-terminator appended. This means that ${html.inline_code(syntax_highlighting.panther("\"\\0\""))} is represented in memory as two null-terminator.`);
+page.text(`As suggested by the type, string literals automatically have a null-terminator appended. This means that ${page.inline_code_block(Language.PANTHER, "\"\\0\"")} is represented in memory as two null-terminator.`);
 page.text("Unlike languages like C/C++ and Zig, nothing special needs to be done to allow string literals to be multi-line.");
 
 
 page.h2("Examples");
-page.code_block(Language.Panther, 
+page.code_block(Language.PANTHER, 
 `// integer literals
 def INTEGER: Int = 12;
 def SCIENTIFIC_INTEGER: USize = 14e2; // 1400