+++ generator/syntax_highlighting/cpp.js
@@ -96,7 +96,7 @@ exports.highlight = function(code){
 
 		}else if(stream.peek() == '/' && stream.peek(1) == '/'){
 			output += "<span class=\"code-gray\">";
-			while(stream.peek() != '\n'){
+			while(stream.EOF() == false && stream.peek() != '\n'){
 				output += stream.next();	
 			}
 			output += "</span>";