+++ generator/syntax_highlighting/llvmir.js
@@ -81,7 +81,7 @@ exports.highlight = function(code){
 
 		}else if(stream.peek() == ';'){
 			output += "<span class=\"code-gray\">";
-			while(stream.peek() != '\n'){
+			while(stream.EOF() == false && stream.peek() != '\n'){
 				output += stream.next();
 			}
 			output += "</span>";